:bricks: separate UnderlineLink from navigation
