+++ src/components/layout/footer/FooterFriends.tsx
