+++ src/components/layout/footer/_data/footer.data.tsx
