+++ src/components/layout/footer/_data/footerTypes.ts
